ai-microservice/requirements.txt
google-generativeai 
python-dotenv
fastapi
uvicorn
python-multipart
PyMuPDF
